agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.12
itsdangerous>=2.1.0
starlette>=0.36.0

langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

markdown>=3.7.0
tabulate>=0.9.0
xhtml2pdf>=0.2.16
aiofiles>=23.0.0
openai>=1.50.0
